# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AI Assistant API
  version: 0.1.0
servers:
  - url: http://localhost:8000
paths:
...
  /v1/tasks/{task_id}/status:
    patch:
      summary: Update task status
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task updated
        "404":
          description: Not found
...